ml/requirements.txt
pandas==2.2.1
paho-mqtt==2.0.0
Faker==24.3.0
scikit-learn==1.4.1.post1
pytz==2024.1
